ultimos ajustes

+++ CrudTarefas.Domain/Aggregates/TarefaAgg/Services/TarefaServices.cs
@@ -1,6 +1,7 @@
 ﻿using CrudTarefas.Domain.Aggregates.Interfaces;
 using CrudTarefas.Domain.Aggregates.Resquests;
 using CrudTarefas.Domain.Aggregates.TarefaAgg.Entities;
+using CrudTarefas.Domain.Common;
 using CrudTarefas.Domain.Interfaces;
 using Microsoft.EntityFrameworkCore;
 
@@ -25,11 +26,23 @@ namespace CrudTarefas.Domain.Aggregates.TarefaAgg.Services
             return add.Entity;
         }
 
-        public async Task Delete(int id)
+        public async Task<ResponseCreateDto<bool>> Delete(int id)
         {
+            var response = new ResponseCreateDto<bool>();
+
             var tarefa = await _unitOfWork.TarefaRepository.GetEntityByIdAsync(id);
+
+            if (tarefa == null)
+            {
+                response.AddWarningValidation("Tarefa inexistente!");
+                return response;
+            }
+
             _unitOfWork.TarefaRepository.Delete(tarefa);
             await _unitOfWork.Commit();
+
+            response.Entity = true;
+            return response;
         }
 
         public async Task<Tarefa?> GetByIdAsync(int id)
@@ -38,16 +51,25 @@ namespace CrudTarefas.Domain.Aggregates.TarefaAgg.Services
         public async Task<List<Tarefa>> GetListAsync()
             => await _unitOfWork.TarefaRepository.GetAll().ToListAsync();
 
-        public async Task<Tarefa> Update(int id, CriarOuAtualizarTarefaRequest request)
+        public async Task<ResponseCreateDto<Tarefa>> Update(int id, CriarOuAtualizarTarefaRequest request)
         {
+            var response = new ResponseCreateDto<Tarefa>();
+
             var tarefa = await _unitOfWork.TarefaRepository.GetEntityByIdAsync(id);
+
+            if (tarefa == null)
+            {
+                response.AddWarningValidation("Tarefa inexistente!");
+                return response;
+            }
 
             tarefa.Atualizar(request.Titulo, request.DataEntrega);
 
             _unitOfWork.TarefaRepository.Update(tarefa);
             await _unitOfWork.Commit();
 
-            return tarefa;
+            response.Entity = tarefa;
+            return response;
         }
     }
 }
